Refactor char_property for further export formats
